change `req` to resolve based on file path rather than directory
We previously had a fabricated file name (`_`) to forcefully resolve a
given directory path to a file path, so `createRequire` would be happy.

Since we now pass the file path every time, we no longer need this.

In future, if we need to `req` a directory, we should resolve the path
at the call-site rather than inside `req`.